Create function to play a round of the game
Created the function, playRockPaperScissors(computerSelection,
playerSelection), that will initiate a round of rock paper scissors.
Testing the choices of the computer and the user against each other
and then determing who wins. This will allow the game to actually be
played.

Also made the CHOICES array a global constant since both functions make
use of it.