fix: prevent too many consecutive round ties
Added logic to the getComputerChoice function, to ensure that if in the
current game, there has been 2 consecutive ties, the cpu will always
choose to either play aggressively or randomly. This stops the cpu
from trying to counter the player's last choice all the time, and
prevents games where there were a ridiculous amount of consecutive
ties.